app/requirements.txt
flask
psycopg2-binary
ultralytics
requests
easyocr
torchvision  
pytesseract